gestion_carteras_api/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
email-validator